backend/pyproject.toml
[tool.poetry]
name = "exampleapp-api"
version = "0.1.0"
description = ""
repository = "https://code.bld.ai/house-numoo/templates/django-react"
authors = ["Ahmed Mohamedeen <ah.mohamadeen@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "4.1.4"
django-environ = "^0.9.0"
celery = "^5.2.7"
psycopg2 = "^2.9.3"
redis = "^4.3.4"
django-celery-beat = "^2.3.0"
django-celery-results = "^2.4.0"
uvicorn = "^0.18.3"
django-cors-headers = "^3.13.0"
djangorestframework = "^3.14.0"
dj-rest-auth = "^4.0.0"
django-allauth = "^0.54.0"

[tool.poetry.group.dev.dependencies]
pylint = "2.15.3"
black = "22.8.0"
pylint-django = "^2.5.3"
pylint-celery = "^0.3"
coverage = "^6.5.0"
safety = "^2.2.0"
django-coverage-plugin = "^2.0.3"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
ensure_newline_before_comments = false
target-version = ["py39"]
include = '\.pyi?$'
exclude = '''
/(
  | \.git
  | \.mypy_cache
  | \.venv
  | dist
  | migrations
)/
'''

[tool.isort]
profile = 'black'
